New PCM model with fragment concept removal and two interrupt sources Renamed size_t/ssize_t to snd_pcm_sframes_t/snd_pcm_uframes_t

+++ src/rawmidi/rawmidi_hw.c
@@ -245,7 +245,7 @@ int snd_rawmidi_hw_open(snd_rawmidi_t **handlep, char *name, int card, int devic
 			snd_ctl_close(ctl);
 			return ret;
 		}
-		if (info.subdevice != subdevice) {
+		if (info.subdevice != (unsigned int) subdevice) {
 			close(fd);
 			goto __again;
 		}
